calculated delta_d (coherency) from fit to parabola
use all data which sit above half height of maximum arranged "evenly" around
maximum. Fit a parabola to these data and use its maximum as d-value. to
calculate FWHH (delta_d) shift the polynomial by -MAX/2 and determine
its roots. the difference of the two roots is FWHH. the coherence length
(length of the periodicity) is proportional to 1/FWHH.

Signed-off-by: Ivo Alxneit <ivo.alxneit@psi.ch>